Upgrade the types of N and Z to avoid possible loss of data
We calculate the value of N by subtracting A from Z. However, A is
stored as uint16_t, but N and Z are uint8_t so the types did not
match. We know that the value of N will fit into a uint8_t but instead
of casting between values, lets just let N and Z have uint16_t types.

+++ include/nuclear-data-reader/nubase_data.hpp
@@ -57,9 +57,9 @@ namespace NUBASE
     /// The mass number
     mutable uint16_t A{ 0 };
     /// The proton number
-    mutable uint8_t Z{ 0 };
+    mutable uint16_t Z{ 0 };
     /// The neutron number
-    mutable uint8_t N{ 0 };
+    mutable uint16_t N{ 0 };
     /// The state level
     mutable uint8_t level{ 0 };
     /// The parity of the spin state
@@ -144,7 +144,7 @@ namespace NUBASE
      *
      * \return Nothing
      */
-    inline void setZ() const { Z = Converter::StringToNum<uint8_t>(full_data, position.START_Z, position.END_Z); }
+    inline void setZ() const { Z = Converter::StringToNum<uint16_t>(full_data, position.START_Z, position.END_Z); }
 
     /**
      * Extract the mass-excess from the NUBASE data file
